Refactor how dialogue text is pulled
Uses the map data to pull dialogue text now, and no longer destroys
colliders once they have been "used". This was causing a cleanup issue.
I could remove them from the list of colliders but switching to a
moveable collider feels better, and might make it easier to set the
color on it instead of the weird jank I did previously.