Make some map modifications to remove weird bug
Because of the distance between the colliders in the road and the
location of the exit to the north. I.e. player's collider is slightly
bigger than a single square and putting them between a single tile space
like that caused the bug. Ignoring the bug for now, because there are
other scaling issues that would need to be addressed to fix it and don't
wanna deal with now :P